complete model
Basically complete the model and simply test it, where SSR means "sum of square of residents". But obviously for complicated dataset, random start model with sigmoid activating function will be easily stuck in local optimum or grad loss  problem. Therefore the next is to apply multiprocessing on main codes to widely search the best optimum and parameters.